Add option to build generated emulation code in the build directory.
This patch adds support for building the generated emu headers in
the build directory instead of the current scheme.  We need it
for our internal build system, and to avoid polluting the source
directory with generated code.

In addition, it speeds up the generation of header files by removing
the string concatenation done during the conversion to hex strings.
In python 2, that can take hours to complete (we still use python 2
in our build system by default).

The patch re-factors the generation of the header files to avoid
repeating common code.

I have tested this by comparing the headers generated.  The results
are identical.