Fix reallocation failures for Twine.
The usage pattern for Twine in these functions seem to trigger a
deallocation that later cause Value::setName to try to allocate more
than UINT32_MAX bytes for the string.

This patch changes the code to generate the Twine objects within the
call to Value::setName.

This was causing std::bad_alloc errors when running 3DMark.  The
benchmark works properly with this patch.