Move readlane's into WWM section to ensure correct reductions.
This change moves the readlanes for subgroup reductions into the WWM
sections. There is a really subtle issue that is super unlikely but
still possible whereby the readlane could be sourced from a register
copy, where the copy happens outwith the WWM section. This would mean
that if it just so happened that the invocation with a
gl_SubgroupInvocationID == 63 was inactive, an undef value could be read
into all lanes.