Ensure that late generated funcs get inlined
Functions llpc.streamoutbuffer.store and llpc.tfbuffer.store get
generated in PatchInOutImportExport.

Not inlining either was causing an assert with a newer LLVM because they
were not being inlined. It looks like an older LLVM would inline them in
the AMDGPU backend, but that is no longer the case in a newer backend.

Fixed by marking them "always inline".

Change-Id: If56f80fca7232c5599efc8e63541ff9cba1cd215