Put pipeline linking into Builder
This is a step towards Builder being the interface between the frontend
and the middle-end.

Also sorted out and documented how Builder is used with individual
shader modules and with a pipeline module.

Includes adding a new class inherited by both BuilderRecorder and
BuilderReplayer that caches the LLVM metadata kind ids used by both.

Change-Id: I4b476cc789970cbb2ed2569a9140d7831d66f68d