Rationalize Builder methods, opcodes and call names
Previously, the opcodes and call names for recorded builder calls were
inconsistent with the builder method names. The method names do not
have a prepended category, which is right, because the division of
BuilderImpl into categories is an internal implementation detail.

This commit similarly changes the opcodes and call names to match --
there is no need for a prepended category, because the categories are an
internal implementation detail of BuilderImpl.

So, for example, to "create a load of a buffer descriptor", the method
name already used that natural order of words:
  Builder::CreateLoadBufferDesc.
I have now changed the opcode to match:
  BuilderRecorder::Opcode::LoadBufferDesc
and I have changed the call name to match:
  llpc.call.load.buffer.desc.

Change-Id: Id37bb3a022d9d8a4c737d27d60d4e69cc3ff6902