Use more unordered maps in EmuLib.
When I was doing some compile time investigation, I noticed that we
spend around 15% of the time searching the symbol table of the archive
for a specific symbol name.  The reason it takes so long is that it is
doing a linear search, and the string comparisons are slow.  I noticed
that this was fix with the latest source.  However, while looking at the
latest code, I found a couple things that could be improved.

Doing searches in std::map data structures where a string is the key can
still take a while.  It is a log n search and the key comparision is not
very fast.  I would suggest changing those to std::unordered_map.  LLVM
already provides a decent hash function for strings.

The other is to avoid duplicating work.  For sets and maps, insert will
search for the element.  If it is not already in the set or map, it will
insert it.  If it is already in, it will not do change the container.
You do not need to check if it is already in the container and then
insert.  That duplicates the search.

These changes do not make a huge difference, but they seem to improve
compile time by about 2% in the pipelines I am looking at.