Integrate amdllpc lit tests into cmake build system
This commit allows the use of "make check-amdllpc" from the normal
driver build directory. It automatically supplies the location of
amdllpc and spvgen.so to the lit tests. The check-amdllpc target depends
on amdllpc and spvgen, so causes them to be built first. It also depends
on the applicable llvm utilities e.g. FileCheck.

Change-Id: I598b0949481616b2c37e79d18d2e20e5a725819e