Reduce optimizations run by llpc
Llpc runs the O3 optimization from llvm-opt.  Many of these
optimizations are focused on optimizing C/C++ code, but are not useful
for code that comes from Vulkan spir-v.

To fix this, I have changed Patch::AddOptimizationPasses to explicityly
list the optimizations it wants to add.  I added the optimizations that
llvm would have added, removing those that would not do anything or seem
redundant.

My analysis of which optimizations are useful was determined by look at
a shaders from games where the shaders where compiled with DXC.  This
could bias the results.

In my tests, the GCN that is output is the exact same, while reducing
the compile time around 5%.

Passes that identified tail calls have been removed.  It changes the
llvm-ir that is generated, but does not change GCN.  Updated the tests
to reflect that change.