Merge user data nodes
The Vulkan driver ensures that the user data nodes for different shader
stages are always compatible. This LLPC commit merges them into a single
set of user data nodes for the whole pipeline, in preparation for a
subsequent commit where the whole pipeline user data nodes will be
passed through the Builder interface.

Change-Id: I2b49a7ba315890a89e381ce5aa5ec55cc8e6129a