Improve codegen for FMin and FMax
The SPIR-V GLSL instructions FMin and FMax say "which operand is the
result is undefined if one of the operands is a NaN", and FClamp is
defined in terms of FMin and FMax.

The SPV_AMD_shader_trinary_minmax instruction FMin3AMD, FMax3AMD and
FMid3AMD say "the result is undefined if one of the operands is a NaN".

In all cases, when implementing them as calls to the llvm.minnum and
llvm.maxnum intrinsics, add the nnan flag to tell the backend that it
doesn't have to worry about NaNs. Without this, the AMDGPU backend
often emits canonicalization code whose only purpose is to convert
signalling NaNs into quiet NaNs so that the intrinsics (which return
the other operand if either operand is a NaN) can be implemented using
the hardware min/max instructions (which return a quiet NaN if either
operand is a signalling NaN).

For example, this IR (without the nnan flags):

define float @maxmin(float %x, float %y, float %z) {
entry:
  %0 = call float @llvm.minnum.f32(float %x, float %y)
  %1 = call float @llvm.maxnum.f32(float %0, float %z)
  ret float %1
}

... gets codegenned into this ISA:

maxmin:                                 ; @maxmin
; %bb.0:                                ; %entry
    s_waitcnt vmcnt(0) expcnt(0) lgkmcnt(0)
    v_mul_f32_e32 v1, 1.0, v1
    v_mul_f32_e32 v0, 1.0, v0
    v_min_f32_e32 v0, v0, v1
    v_mul_f32_e32 v1, 1.0, v2
    v_max_f32_e32 v0, v0, v1
    s_setpc_b64 s[30:31]
.Lfunc_end0:

Adding the nnan flag removes the multiplies by 1.0, which are only
there to canonicalize NaNs.