Separated descriptor load into three parts
This commit separates a descriptor load into three parts:
1. Get pointer to descriptor
2. Optionally index it
3. Load the descriptor from the pointer.

This separation is done both in the builder interface and in the
llpc.descriptor.* calls lowered by PatchDescriptorLoad.

This only affects image-related descriptors (image, fmask, texel buffer,
fmask), not buffer descriptors.

I have done this to allow a future commit to change the SPIR-V reader to
use builder calls directly, instead of needing the LowerImageOp pass
afterwards. In most cases the SPIR-V reader would be able to trace back
from an image op to its descriptor's set, binding and index. But, in the
corner case of a pointer to (array of) descriptor being passed as a
function parameter, it is not possible for the SPIR-V reader to do that.

The implementation in PatchDescriptorLoad currently assumes that
everything is inlined by then, so it can trace back from the "descriptor
load from pointer" back to the "get pointer to descriptor", and
implement them together, including the case of optimizing a "dynamic
descriptor" (one that is in the top-level user data) or an immutable
sampler (constant data provided to the compiler). The plan is to remove
this assumption of inlining everything in a future commit.

Change-Id: I1275ecc1774afca4d7361479445cee0b900852f5