Builder trinary minmax
This adds only one new builder method, for fmed3, implemented as the
fmed3 instruction for float, or a combination of fmin and fmax for half.

The SPIR-V reader has been changed to use that builder method, and for
all other trinary_minmax instructions it just generates IR itself using
standard LLVM IR instructions and intrinsics.

Change-Id: If0f3d9e54c6ff79fc1a13871f817bb9eeccb0465