Update shaderdb tests
Update the following shaderdb test:

* OpImageQuerySizeLod_TestTextureSize_lit.frag
* OpImageQuerySize_TestTextureSize_lit.frag
* OpImageQuerySize_TestBasic_lit.frag

Make sure that all returned values of textureSize() will be used
in the shader output value, and that value is not undef.

* OpImageQuerySize_TestBasic_lit.frag
* OpImageQuerySize_TestImage_lit.comp

Be more flexible in checking an intrinsic type after patching.
Currently in two cases the type of @llvm.amdgcn.image.getresinfo.cube
to check against is a 4-element vector, but can theoretically be
a 3-element one.