Consistently set FMF in builder
This changes the SPIR-V reader to consistently set the fast math flags
in Builder when generating an instruction.

Code that uses non-builder functions to generate IR, or uses a builder
function that does not honor FMF (e.g. CreateMinNum), needs to call
SPIRVToLLVM::setFastMathFlags to set the FMF on the instruction from the
builder.

However, in this commit, I did not want to introduce any actual
generated code changes, so in several places I also added a
clearFastMathFlags and a TODO: that we need to remove it and fix up the
tests. That should happen in future commits.

Change-Id: I0a2d7180de101eee2c6f8ee424a7976a1ccda68e