Reunited pre-patch and patch pass managers
Part of the drive to clean up the interface from the front-end into the
middle-end.

Ultimately we want the whole of the middle- and back-ends to run in the
same pass manager. This is the first half of that.

The change includes factoring out shader cache checking in the graphics
pipeline. That remains in the front-end, in a lambda that is passed in
to the patch phase to run part way through it.

Change-Id: I1b71c042f2f4c95d2b24510143aaaf730492f019