BuilderContext and other middle-end interface changes
Part of a series of patches to clean up the middle-end interface.

The middle-end interface now consists of the following classes:
 - BuilderContext: holds state across multiple compiles; used to create
   Pipeline and Builder;
 - Pipeline: in a pipeline compile, holds pipeline state, and has
   methods for linking the pipeline IR module and then running middle- and
   back-end passes to generate to ISA;
 - Builder: used by the front-end to create IR, with specialist methods
   for graphics constructs.
 - PassManager: subclass of legacy::PassManager with a couple of extra
   features.

Pipeline is an interface class for PipelineState, which is visible only
in the middle-end. PassManager is an interface class for
PassManagerImpl.

See the outline in LgcOverview.md for more details.

Currently the only state held by Pipeline is the user data nodes.
I will put other state into it in subsequent comments.

Change-Id: Ia31c09715706cfc3165bd0c4b750c184050ea6e8