Created BuilderContext class
Part of the drive to clean up the interface from the front-end into the
middle-end.

The BuilderContext class will keep state shared by different builders
from different compiles. Soon it will hold the target GPU; later on we
could use it to share the middle-end pass manager so we do not have to
create a new one every compile.

The front-end now uses BuilderContext to create a Builder.

Also restructured the comment at the top of llpcBuilder.h to be a list
of actions the front-end takes to use the middle-end.

Change-Id: I1b9c6886d7b0392d351d93b0ff3bdee29adaa151