PipelineState allocation changes
Part of the drive to clean up the interface from the front-end into the
middle-end.

Builder now only allocates a PipelineState if it was not given one via
Builder::SetPipelineState. The inner BuilderImpl created by
BuilderReplayer now has the outer PipelineState passed in to it.

This commit also manually adds PipelineStateWrapper to the middle-end
pass manager so it can give it the PipelineState it already has.

Change-Id: I538aa83a07e0eb230e3c26629958fd84528b806a