Group BuiltInTessLevel* stores together
Make the code of BuiltInTessLevelInner and BuiltInTessLevelOuter
more merge-friendly, by grouping stores generated in WriteValueToLds
at the end of block and keeping stores generated in StoreTessFactorToBuffer
where they were.

This helps in a common case, where stores to BuiltInTessLevelOuter (or
BuiltInTessLevelInner) are listed sequentially in the source, and the
generated instructions are intermingled.