llpcInternal: Refactor EmitCall in terms of the IRBuilder<> interface
Add a new variant of the EmitCall helper that takes an IRBuilder<>
reference, which is conveniently usable in all contexts that use a builder
and avoids the need to manually distinguish between an insertion point
in the middle vs. at the end of (an incomplete) basic block.

Further benefits of this refactoring include:

- Use a SmallVector for argTys to avoid a heap allocation in most cases.
- Use the builder interface for creating a function call instead of the
  manual Create method; this is the more idiomatic way.
- Do not set call instruction attributes redundantly.

v2: fix up lit tests