Prepare for upstream LLVM changes in LLPC
Changes in upstream LLVM require preparation in LLPC. The changes here are
compatible with both the current version of LLVM (to ensure that it still
continues to build and work), as well as the new scheme.

It is possible to implement the alignment changes more succinctly, but not in a
way that's compatible with current and future LLVM changes.