Created ShaderModes to pass input language shader modes into middle-end
Part of a series of patches to clean up the middle-end interface.

ShaderModes is a new middle-end object for storing shader modes (FP
mode, workgroup size, tessellation modes, geometry modes, fragment
modes) that the front-end reads from its input language.

For a pipeline compile, ShaderModes is embedded in PipelineState,
and the latter ensures that the former is recorded in IR metadata when
recording.

For a shader compile, there is no PipelineState, so BuilderRecorder
creates its own ShaderModes. The front-end must call
Builder::RecordShaderModes to ensure it is recorded in IR metadata.

For each part of the shader modes:

* The SPIR-V reader calls each Builder::Set*ShaderMode to set the modes
  for a shader, depending on what shader stage it is.

* Middle-end code that needs to access it calls
  PipelineState::GetShaderModes to get the ShaderModes object,
  then calls Get*ShaderMode on that to get the particular part of the
  modes.

The following shader modes use this mechanism:

* Common shader mode (the FP mode) for all shader stages.
  - We still keep the original FP mode flags from the SPIR-V in the
    front-end as well (although no longer in ResUsage), because they are
    used by the front-end arith optimization pass.
  - Also commoned up the two identical ConfigBuilder functions to create
    the FP mode register settings.

* Tessellation mode. This gets set for both TCS and TES, and the
  middle-end merges the settings together.
  - vertexSpacing, vertexOrder, primitiveMode, pointMode, outputVertices

* Geometry mode for GS.
  - inputPrimitive, outputPrimitive, invocations, outputVertices

* Fragment mode for FS.
  - pixelCenterInteger, earlyFragmentTests, postDepthCoverage

* Compute mode for CS, containing the workgroup size.
  - Also moved the workgroup reconfigure optimization into
    PatchInOutImportExport, as it sets a ResUsage.
  - Also removed unused code from the same place to detect use of
    ViewIndex.

Change-Id: Icf7cb30dfa44145e9c25cb6931e5f3947369ef4b