-emit-lgc option
This causes amdllpc to output LLVM IR assembly suitable for input to the
lgc utility.

Using -emit-lgc causes BuilderRecorder to omit the opcode metadata that
it usually adds to each lgc.create.* function declaration. That is so
that LGC tests generated with -emit-lgc rely on the more stable
lgc.create.* name rather than the less stable opcode.

Change-Id: I13a2c3a5f0374ba2d71561354d38de31a5a48bce