Use more literal struct types instead of identified struct types
Identified struct types are by definition not uniqued, even if you try
to create another one with the same name and the same members, so
creating lots of them unnecessarily is wasteful.

You can see this in the output of "amdllpc -print-after-all", where the
initial IR sometimes has hundreds of structs like this:

%llpc.matrix.row.220 = type <{ [4 x float] }>
%llpc.matrix.row.221 = type <{ [4 x float] }>
%llpc.matrix.row.222 = type <{ [4 x float] }>

... which are all identical except that LLVM has been forced to pick a
suffix for each name just to make them different.