TargetMachine is not dependent on pipeline options
Part of a series of patches to clean up the middle-end interface.

The TargetMachine is not dependent on pipeline options, so this commit
removes the mechanism to remember the pipeline options from the last
TargetMachine and create a new one if they have changed. Now the
TargetMachine is created just once for the Context.

Change-Id: Ief72a428271cb3c83c60be5840b4a36baba4ea3b