Stop using GetTargetMachinePipelineOptions in front-end
Part of a series of patches to clean up the middle-end interface.

The front-end has a couple of places where it uses pipeline options
(scalarBlockLayout and robustBufferAcces). This is wrong, and should be
fixed. But for now this commit handles these two options specially with
methods in Context, keeping the existing behavior in a shader compile of
remembering the last value from the last pipeline compile.

This removes all front-end use of GetTargetMachinePipelineOptions, which
then means a subsequent commit can remove it, because in the middle-end
we can access the pipeline options more directly.

Change-Id: Ibd8ef80647c2c0cc6cd66757180530f14b9afd92