Move ResourceUsage and InterfaceData into middle-end
Part of a series of patches to clean up the middle-end interface.

These per-shader structs now live in PipelineState and are defined in
their own header file llpcResourceUsage.h, and are private to the
middle-end.

Also moved a couple of small structs only used in the middle-end into
llpcResourceUsage.h.

Change-Id: I890b005f4a629239cfddcbd52412eeccd93d2d0e