Removed uses of spv:: built-in IDs in patch
Part of a series of patches to clean up the middle-end interface.

Any use of a built-in ID in builder and patch is now one of the
BuiltInKind values defined in builder.

Also rearranged header files so a source file in patch can #include
llpcBuilderBuiltIns.h to see those definitions.

Change-Id: I77bc8d31653cd179bff7793cfd7f6d69df7d825f