Started LlpcOverview.md and LgcOverview.md documents
Part of a series of patches to clean up the middle-end interface.

These new files outline the state of LLPC and LGC (the middle-end part
of LLPC) as they will be at the end of the "middle-end interface
clean-up" series of patches.

LgcOverview.md contains the material that was in the big comment at the
top of Builder.h.

Change-Id: Ie1b21f5ddbba597c80c5d833d8abe0c9685a6c36