llpc: Document policy of required zero-initialization of structures
The intent of this policy is to allow us to bump the interface major
version less frequently, and to avoid having to guard lots of code with
preprocessor macros that depend on the interface version.

Such guarding tends to be fragile because it explodes the matrix of
possible compilation options.

v2: Stylistic change