Misc changes ready for coding standard change
- Removed m_ prefix from struct fields in config builder. clang-tidy
can't cope with changing them where they are used in macros.
- In llpcPipelineDumper.cpp and llpcUtil.cpp, pre-changed variable
pString to string as its name is hard-coded into macros, which
clang-tidy identifier renaming cannot cope with.
- In SPIRVReader, removed macro HANDLE_OPCODE that clang-tidy identifier
renaming cannot cope with.
- In llpcPipelineState, vkgcPipelineDumper and vkgcUtil, change pString
to string as it is used in a macro, which fools clang-tidy.
- In config builders, change containsPalAbiMetadataOnly to CamelCase as
it is a const member, and its use in a template fools clang-tidy.

Change-Id: I92224fbb6b9b654e37d7a89b6c49da4f8e716762