Memory leak in llpc
1.There are memory leak during vkCreateModule.
The root cause is that, when LLpc call Compiler::BuildShaderModule, it will create new array to store debug info trimed binary code,
and assign the pointer to the moduleDataEx, but after that, when copy the pCode from moduleDataEx to pModuleDataEx, our driver forget to release the pointer of moduleDataEx,
As result, the memory leaked.

2.When call ~Compiler(), the "return" will cause ShaderCache not been
destory.