Moved most of LLPC into llpc subdirectory
Future commits will move some code out of the llpc subdirectory that is
not strictly part of LLPC, such as the pipeline dumper.

This commit leaves a test/CMakeLists.txt that forwards to
llpc/test/CMakeLists.txt, to cope with an old build script that still
looks for it in the old location.

Change-Id: I2db61847817553653c400917b3745cc50cf58f5f