Bound the number of InstCombine iterations
By default, InstCombine iterates over a function until it keeps
changing. This causes the last iteration to always go to waste.
It is offen possible to predict the maximum number of itrations
necessary to reach a fixpoint and not run the additional iteration that
would detect it.

This patch bounds InstCombine passes in llpPatch and spirvLower with
conservatives numbers that should have minimal or no effect on the
generated code. On a database on 7400 real-world pipelines, this saves
about 4% of of InstCombine iterations. In the future it should be
possible to shave off more compilation times by making the thresholds
tighter at the expense of optimizing code less.