Prepare for additional immarg string appearing in arguments
A change in llvm means that a couple of arguments to calls previously looking
like "i32 2" will become "i32 immarg 2" - which is in line with other immediate
arguments.

In order to cope with the transition, a regular expression is added which means
that FileCheck will match both options. The regular expression can be removed
once llvm changes have been merged - although the change is benign.