Add option to amdllpc to build shader variants.
Part of https://github.com/GPUOpen-Drivers/llpc/issues/431.

This PR introduces the options that will be needed to use amdllpc as a
cache creator tool.  The intended usage would be something like:

```
amdllpc -auto-layout-desc -gfxip=9.0.0 --use-relocatable-shader-elf -shader-cache-mode=2 -shader-cache-filename=$SHADER_CACHE_FILENAME -build-shader-cache <list of spv files>
```

In the future, we may have to add other options as the tool will require
more input, but this would be the initial implementation.