Add pass to simplify intrinsics
This is a rebase of the original patch from Neil Henning.

It adds folding for imageLoads and texture samples where it can detect
the source of the indices would fit in a16 addressing, and replaces
the original load/sample with a16 variants.

It also detects sin(x / 2pi) / cos (x / 2pi) and replaces the LLVM
intrinsic with the amdgcn native one (saving us 2 instructions per call).