llpcCompiler: Simplify preprocessor directives around cache-related methods
Interface version 38 removed the application shader cache that is passed to
the compiler as part of the pipeline build info, but that doesn't mean we
have to make code harder to read using preprocessor directives.

Unify the two paths by just keeping the application cache pointer as a
nullptr when it cannot exist.

v2:
- Renaming and minor code style adjustment
- Adjust how LookUpShaderCaches builds the shader cache array