Separate llpcDebug.h into front-end and middle-end parts
Part of formalizing the split between front-end and middle-end.

This commit adds a new llpcBuilderDebug.h for use by the middle-end,
leaving llpcDebug.h for use by the front-end.

Each contains a definition of LLPC_OUTS. The middle-end one in
llpcBuilderDebug.h uses a new implementation in which the front-end
calls the middle-end to tell it a raw_ostream to use for LLPC_OUTS,
defaulting to no output.

Change-Id: Iac2f6a04bbe6de091be8bb61eb51e1d9df68feef