Fixed llpcElfWriter to not depend on middle-end register defs
Part of formalizing the split between front-end and middle-end.

llpcElfWriter is in the front-end, so we do not want it using non-public
parts of the middle-end. (If we were to move the linking and ELF-writing
code into the middle-end one day, then we would probably implement it
using existing LLVM library code.)

Change-Id: Ibf38d2586d9aed083e2121602e0cf981bb690cbb