Separate llpcUtil and llpcInternal
Part of formalizing the split between front-end and middle-end.

This commit finishes the separation of llpcUtil (used by the front-end)
and llpcInternal (used by the middle-end).

Also added llpcSpirvLowerUtil, to add copies of
GetShaderStageAbbreviation and GetResourceMappingNodeTypeName for use by
the front-end, and some other front-end-only defines.

Also some other changes to #includes for front-end/back-end separation.

Change-Id: Id4c39287a1751dbc609a6077c0faba1663f62b24