Moved partial pipeline compile const data check into middle-end
Part of formalizing the split between front-end and middle-end.

We can only use partial pipeline compilation, merging two pipeline ELFs,
if there is no constant data in any shader other than the FS.

The check for this was in front-end code. This commit moves the check to
middle-end, as it uses middle-end information.

Change-Id: Ie8cc6c01554fa17178f4491dff948394a27a2e58