Added lgc command-line utility
It takes an LLVM IR module as output by amdllpc -emit-lgc, and runs it
through LGC (the LLPC middle-end) and LLVM back-end.

It allows multiple IR modules in the same source file, to make it
possible to put multiple related small tests into the same file.

It has an -extract option to only compile one of multiple IR modules in
the source file, to make it easier to write a test that compiles two
shaders separately then links them.

Change-Id: I1bcea68e68da42ec4d2de568a7e231a010c6305c