Removed PAL_CLIENT_INTERFACE_MAJOR_VERSION from LGC
Part of preparing for LGC to be separated into its own directory.

Where some code or metadata generation needs to depend on the PAL ABI
version, LGC now uses the PAL ABI version passed in to
BuilderContext::Create by the front-end. The lgc standalone utility now
has a -pal-abi-version option to set it, defaulting to the latest known.

Change-Id: I379deabc8966229e575200ac35240f8b7790a96e