Overhaul converting sampler finding code
This modifies the code that spots that a sampler is a YCbCr converting
sampler, allowing us to switch Builder's idea of the type of a sampler
back to v4i32.

The design of the Vulkan YCbCr-converting extension is such that you
can't tell in the SPIR-V that a sampler is a converting sampler, so the
use of a sampler needs to know the descriptor set and binding to know
how to treat it. That conflicts with SPIR-V's ability to pass a pointer
to an image or sampler descriptor around through a call arg or a phi
node (although the latter may just be an artifact of LLVM
optimizations).

Before this commit, the detection code used to be in BuilderImplDesc at
the point that the Builder::CreateLoadDescFromPointer is implemented, by
scanning back to the llpc.descriptor.get.sampler.ptr. If it turned out
to be a converting sampler, the immutable value was used as the result
of Builder::CreateLoadDescFromPointer. A problem with this approach was
that the Builder interface, and the IR it generates, had to pretend that
a sampler is v8i32 instead of v4i32.

With this commit, a sampler is back to being v4i32 in the IR, and the
detection of a converting sampler by scanning back is done at the point
of use, when implementing Builder::CreateImageSample.

Change-Id: I3d775746549553f38b331669d58e87d3eed3079c