Revamped PatchDescriptorLoad
This commit changes the code to implement image and sampler descriptor
loading with the semantics intended by the separate Builder calls to
1. get the descriptor pointer
2. add an index to a descriptor pointer
3. load the descriptor from the pointer

Before this commit, the code relied on being able to scan back from (3)
to (1) to discover the descriptor set and binding. If we have
non-inlined subroutines in the future, that is no longer possible, as
SPIR-V allows the passing of a descriptor pointer as a call arg.

An immutable sampler is now handled by putting it into a read-only
global variable at (1), and relying on LLVM optimizations to promote it
back to a constant in the common case.

Change-Id: Ic37fcb2858dfbc0a84b065dd7987bd4a2746e0e3