More conservative handling of degenerate OpMemoryBarrier instructions
Before PR #477, an OpMemoryBarrier that didn't set any ordering bits
would be translated to an LLVM IR fence instruction with no ordering,
which is invalid IR.

PR #477 changed the translator to ignore such OpMemoryBarrier
instructions, treating them as a no-op.

Unfortunately some games rely on getting some ordering from an
OpMemoryBarrier instruction with the semantics set to e.g. 0x200, which
includes the CrossWorkgroupMemory flag but none of the ordering bits.
Handle this by applying a default ordering of AcquireRelease.