Separated GfxIpVersion into vkgc and lgc versions
Part of formalizing the split between front-end and middle-end.

I decided that creating a file lgc/Defs.h for definitions shared by
XGL/LLPC on one hand and LGC on the other hand wasn't really the right
approach, and layering will be cleaner if they have separate
definitions.

This commit makes separate front-end and middle-end definitions of
GfxIpVersion. It is not used in the LGC interface.

Change-Id: I4d5a1c093ead0f7d060245345ed82717b33b1185