Separated MaxColorTargets definitions
Part of formalizing the split between front-end and middle-end.

I decided that creating a file lgc/Defs.h for definitions shared by
XGL/LLPC on one hand and LGC on the other hand wasn't really the right
approach, and layering will be cleaner if they have separate
definitions.

This commit creates separate definitions for MaxColorTargets. It is not
part of the LGC interface.

Change-Id: I0cc8fb932d00363839aa937eb52b063dc54fe9fb