Separate ResourceMappingNode definition
Part of formalizing the split between front-end and middle-end.

I decided that creating a file lgc/Defs.h for definitions shared by
XGL/LLPC on one hand and LGC on the other hand wasn't really the right
approach, and layering will be cleaner if they have separate
definitions.

There was already a a separate ResourceNode definition for internal LGC
use, but the LGC interface was using LLPC's ResourceMappingNode and
ResourceMappingNodeType. This commit changes the LGC interface to use
LGC's ResourceNode and a new ResourceNodeType.

Change-Id: Ia3afc7fbaa85a3eabf135fca40ddc6b2d418d279