Separated BasicType definition
Part of formalizing the split between front-end and middle-end.

I decided that creating a file lgc/Defs.h for definitions shared by
XGL/LLPC on one hand and LGC on the other hand wasn't really the right
approach, and layering will be cleaner if they have separate
definitions.

This commit creates separate LLPC and LGC definitions of BasicType. It
is not part of the LGC interface.

Change-Id: Ib5fb86d16bf45c3842653408310e1dab3000405b