Remove now empty lgc/Defs.h
Part of formalizing the split between front-end and middle-end.

I decided that creating a file lgc/Defs.h for definitions shared by
XGL/LLPC on one hand and LGC on the other hand wasn't really the right
approach, and layering will be cleaner if they have separate
definitions.

This commit is the last of the series in this PR. It removes the now
empty lgc/Defs.h.

Change-Id: I7628d656bdc9a0f4da342ca5e0887f19bc94fc88