Moved GetLgcShaderStage to avoid header file problem
When I moved LGC into its own lgc directory, I hit a problem that
llpcUtil.h needs to include llpcBuilderCommon.h, because the former
declares GetLgcShaderStage which uses lgc::ShaderStage. But non-LLPC
uses of llpcUtil.h do not have the new location of llpcBuilderCommon.h
on the include path.

To work around that, I moved GetLgcShaderStage to a different place.

Change-Id: I85ccaedcad1f0ec04ec7ab14f61e3f116fb8b1b4