Add pipeline options for shadow descriptor table.
This allows the VA high ptr value to vary on a per-pipeline basis.
Old command line options are still present and override pipeline
options.

Specifically this is required to address a problem exposed by
PR503 which stopped ignoring variations in VA high ptr value
supplied on the command line.

The pipeline option to enable the shadow descriptor table is an
integer, with 0 meaning auto (use defaults).  This means a zero
initialised pipeline options structure will not override the
default behaviour.