Re-instate direct sgpr use for top-level compact descriptor
Part of my change "Revamped PatchDescriptorLoad" was to handle a
descriptor in the top-level table (used for a Vulkan "dynamic
descriptor") by s_loading it from the spill table, to make handling it
more uniform with other descriptors.

In the case of a compact buffer descriptor, that caused a performance
regression, because the previous code would pass the two-dword
descriptor as part of user data sgprs.

This commit reinstates that behavior for a compact buffer descriptor in
the top-level table.

Change-Id: I46e4fca9a440aeb9c51a52ac618081b616988b40