llpcBatchBufferOp: Downgrade -Wdeprecated-declarations to a warning
There is a pending LLVM change that deprecates getParamAlignment in
favor of getParamAlign, using the new type-safe MaybeAlign wrapper.

Adding the LLVM change causes some CI builds to fail with a deprecation
warning-turned-error due to -Werror. This is mitigated by this commit.
Since getParamAlign is introduced in the same LLVM commit as the
deprecation, there is unfortunately no way to properly fix LLPC
ahead of time: instead, we'll have to change the uses of
getParamAlignment to getParamAlign once the LLVM commit has propagated
to all relevant locations.