Changed llpc.* funcs to lgc.* funcs
LGC inserts and later lowers a number of function declarations and
calls. This commit changes their names to have an lgc prefix instead of
an llpc prefix. It also changes the BuilderRecorder prefix from
llpc.call to lgc.create (to reflect the fact that they come from
Builder::Create* calls).

This commit has minimal code changes but edits to all the lit tests.

Change-Id: I531eb7af4c5ac35aa7681a98b1d6586a876aeec6